requirements.txt
python-telegram-bot>=20.7,<22
requests>=2.31,<3
beautifulsoup4>=4.12,<5
lxml>=4.9,<6
python-dotenv>=1.0,<2
playwright>=1.41,<2
